Moved OCRs from ocr.py to their respective files
