+++ ocr/model_32px.py
@@ -1,10 +1,127 @@
 import math
 from typing import List
 from collections import defaultdict
+import os
+import shutil
+import cv2
+import numpy as np
+import einops
 
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
+
+from utils import Quadrilateral, chunks
+from .common import OfflineOCR
+from .model_48px_ctc import generate_text_direction
+
+class Model32pxOCR(OfflineOCR):
+    _MODEL_MAPPING = {
+        'model': {
+            'url': 'https://github.com/zyddnys/manga-image-translator/releases/download/beta-0.3/ocr.ckpt',
+            'hash': 'd9f619a9dccce8ce88357d1b17d25f07806f225c033ea42c64e86c45446cfe71',
+            'file': '.',
+        },
+    }
+
+    def __init__(self, *args, **kwargs):
+        os.makedirs(self._MODEL_DIR, exist_ok=True)
+        if os.path.exists('ocr.ckpt'):
+            shutil.move('ocr.ckpt', self._get_file_path('ocr.ckpt'))
+        super().__init__(*args, **kwargs)
+
+    async def _load(self, device: str):
+        with open('alphabet-all-v5.txt', 'r', encoding = 'utf-8') as fp:
+            dictionary = [s[:-1] for s in fp.readlines()]
+
+        self.model = OCR(dictionary, 768)
+        sd = torch.load(self._get_file_path('ocr.ckpt'), map_location = 'cpu')
+        self.model.load_state_dict(sd['model'] if 'model' in sd else sd)
+        self.model.eval()
+        self.use_cuda = device == 'cuda'
+        if self.use_cuda:
+            self.model = self.model.cuda()
+
+    async def _unload(self):
+        del self.model
+    
+    async def _forward(self, image: np.ndarray, textlines: List[Quadrilateral], verbose: bool = False) -> List[Quadrilateral]:
+        text_height = 32
+        max_chunk_size = 16
+
+        quadrilaterals = list(generate_text_direction(textlines))
+        regions = [q.get_transformed_region(image, d, text_height) for q, d in quadrilaterals]
+        out_regions = []
+
+        perm = range(len(regions))
+        if len(quadrilaterals) > 0 and isinstance(quadrilaterals[0][0], Quadrilateral):
+            perm = sorted(range(len(regions)), key = lambda x: regions[x].shape[1])
+
+        ix = 0
+        for indices in chunks(perm, max_chunk_size):
+            N = len(indices)
+            widths = [regions[i].shape[1] for i in indices]
+            max_width = 4 * (max(widths) + 7) // 4
+            region = np.zeros((N, text_height, max_width, 3), dtype = np.uint8)
+            for i, idx in enumerate(indices):
+                W = regions[idx].shape[1]
+                region[i, :, : W, :] = regions[idx]
+                if verbose:
+                    os.makedirs('result/ocrs/', exist_ok=True)
+                    if quadrilaterals[idx][1] == 'v':
+                        cv2.imwrite(f'result/ocrs/{ix}.png', cv2.rotate(cv2.cvtColor(region[i, :, :, :], cv2.COLOR_RGB2BGR), cv2.ROTATE_90_CLOCKWISE))
+                    else:
+                        cv2.imwrite(f'result/ocrs/{ix}.png', cv2.cvtColor(region[i, :, :, :], cv2.COLOR_RGB2BGR))
+                ix += 1
+            image_tensor = (torch.from_numpy(region).float() - 127.5) / 127.5
+            image_tensor = einops.rearrange(image_tensor, 'N H W C -> N C H W')
+            if self.use_cuda:
+                image_tensor = image_tensor.cuda()
+            with torch.no_grad():
+                ret = self.model.infer_beam_batch(image_tensor, widths, beams_k = 5, max_seq_length = 255)
+            for i, (pred_chars_index, prob, fr, fg, fb, br, bg, bb) in enumerate(ret):
+                if prob < 0.7:
+                    continue
+                fr = (torch.clip(fr.view(-1), 0, 1).mean() * 255).long().item()
+                fg = (torch.clip(fg.view(-1), 0, 1).mean() * 255).long().item()
+                fb = (torch.clip(fb.view(-1), 0, 1).mean() * 255).long().item()
+                br = (torch.clip(br.view(-1), 0, 1).mean() * 255).long().item()
+                bg = (torch.clip(bg.view(-1), 0, 1).mean() * 255).long().item()
+                bb = (torch.clip(bb.view(-1), 0, 1).mean() * 255).long().item()
+                seq = []
+                for chid in pred_chars_index:
+                    ch = self.model.dictionary[chid]
+                    if ch == '<S>':
+                        continue
+                    if ch == '</S>':
+                        break
+                    if ch == '<SP>':
+                        ch = ' '
+                    seq.append(ch)
+                txt = ''.join(seq)
+                print(prob, txt, f'fg: ({fr}, {fg}, {fb})', f'bg: ({br}, {bg}, {bb})')
+                cur_region = quadrilaterals[indices[i]][0]
+                if isinstance(cur_region, Quadrilateral):
+                    cur_region.text = txt
+                    cur_region.prob = prob
+                    cur_region.fg_r = fr
+                    cur_region.fg_g = fg
+                    cur_region.fg_b = fb
+                    cur_region.bg_r = br
+                    cur_region.bg_g = bg
+                    cur_region.bg_b = bb
+                else:
+                    cur_region.text.append(txt)
+                    cur_region.fg_r += fr
+                    cur_region.fg_g += fg
+                    cur_region.fg_b += fb
+                    cur_region.bg_r += br
+                    cur_region.bg_g += bg
+                    cur_region.bg_b += bb
+
+                out_regions.append(cur_region)
+        return out_regions
+
 
 class ResNet(nn.Module):
